requirements.txt
discord.py[voice]>=2.5.2
pyyaml>=6.0.2